+++ examples/components/index.ts
@@ -1,0 +1,4 @@
+export * from './Header';
+export * from './Example';
+export * from './Main';
+export * from './LanguageSwitcher';