+++ examples/pages/index.ts
@@ -1,0 +1,3 @@
+export * from './HttpBackend';
+export * from './Simple';
+import './styles.css';